remove unneccessary type annotations, split cell into separate file, add empty input tests
