+++ src/main/scala/au/com/cba/omnia/grimlock/spark/examples/PipelineDataPreparation.scala
@@ -165,7 +165,7 @@ object PipelineDataPreparation {
 
       val csb = d
         .slice(Over(Second), rem2, false)
-        .transformWithValue(transforms, stats.toMap(Over[Position2D, Dimension.First](First)))
+        .transformWithValue(transforms, stats.toMap(Over(First)))
         .slice(Over(Second), rem3, false)
 
       (ind ++ csb)
